+++ R/typeR-package.R
@@ -1,0 +1,8 @@
+#' @keywords internal
+#' @importFrom stats coef rnorm
+#' @importFrom utils flush.console head
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL